% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_dist_standardised.R
\name{degree_dist_standardised}
\alias{degree_dist_standardised}
\title{Degree distribution standardised}
\usage{
degree_dist_standardised(g, maxk = 1000)
}
\description{
Degree distribution standardised
}
