% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_net.R
\name{KL_net}
\alias{KL_net}
\title{KL Divergence of mechanistic and statistical network models}
\usage{
KL_net(theta_m, n = 10, theta_s = 3, replicates = 1000,
  sorted = TRUE, mech_net, mech_args, lstat, lapply_opt = TRUE,
  ds_return = FALSE, pmf = NULL, mirror = TRUE)
}
\arguments{
\item{theta_m}{numeric mechanistic model parameters}

\item{n}{integer number of nodes}

\item{theta_s}{numeric statistical model parameters}

\item{replicates}{numeric number of replicates}

\item{sorted}{logical whether to consider the sorted or unsorted degree sequence as the integral statistic}

\item{mech_net}{function the mechanistic network simulator}

\item{mech_args}{list arguments to the mechanistic network simulator}

\item{lstat}{function computes the likelihood statistic}
}
\description{
KL Divergence of mechanistic and statistical network models
}
\examples{

theta_m = 5
theta_s = 1
n = 15

mech_net = function(m, n, args){args$m = m; args$n = n; do.call(igraph::sample_pa, args)}
mech_args = list(directed = FALSE)
lstat = function(x){sum(igraph::degree(x) == 6)}

KL_net(
  theta_m = theta_m,
  theta_s = theta_s,
  n = n,
  mech_net = mech_net,
  mech_args = mech_args,
  lstat = lstat
)

}
