% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_net.R
\name{net_ss}
\alias{net_ss}
\title{KL Divergence of mechanistic and statistical network models}
\usage{
net_ss(theta_m, replicates = 1000, sorted = TRUE, mech_net, lstat,
  lapply_opt = TRUE)
}
\arguments{
\item{theta_m}{numeric mechanistic model parameters}

\item{replicates}{numeric number of replicates}

\item{sorted}{logical whether to consider the sorted or unsorted degree sequence as the integral statistic}

\item{mech_net}{function the mechanistic network simulator}

\item{lstat}{function computes the likelihood statistic}

\item{lapply_opt}{boolean internal}
}
\description{
KL Divergence of mechanistic and statistical network models
}
\examples{

theta_m = 5
theta_s = 1
n = 15

mech_net = function(m, n, args){args$m = m; args$n = n; do.call(igraph::sample_pa, args)}
mech_args = list(directed = FALSE)
lstat = function(x){sum(igraph::degree(x) == 6)}

KL_net(
  theta_m = theta_m,
  theta_s = theta_s,
  n = n,
  mech_net = mech_net,
  mech_args = mech_args,
  lstat = lstat
)

}
